@@ -1,10 +1,10 @@
 name: e_notification_platform_interface
 description: Easy to Use Notification Platform Interface
-version: 0.4.0
+version: 1.0.0
 homepage: https://iamee.me
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.12.0"
 
 dependencies:
